requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-oauth-toolkit==1.7.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8

requests==2.31.0
